requirements.txt
requests>2
robotframework>3
pyOpenSSL>20
requests_oauthlib>1
pyyaml>5
jsonschema>3